Update printReceipt() to print inorder and add different print forms
This commit updates the printReceipt() method to function using the new
Item data structure. A Queue has also been introduced to keep track of
the order in which items are scanned and hence enables them to be
printed in that order. A number of methods have also been created to
enable items to be printed in various formats including with the price
first.