Update JUnit tests
This commit adds new unit tests to test the new features as well as
modifying existing JUnit tests to simplify them. A setUp() and
tearDown() has been added to enable objects and variables to be created
and destroyed in an organised manner. This refactoring will mean that
maintaining and updating the tests going forward should be relatively
straightforward.